requirements.txt
flask
flask-login
python-dotenv
openai
stripe>=10.0.0
PyPDF2
requests
beautifulsoup4
